Add files via upload
This code/webpage is useful for apartments who want their residents to update the database through their computers instead of having a supervisor or any other person update the data. It takes the FirstName, LastName, Receipt Number From The Receipt Given By The Person Responsible For Receipt Distribution & Month Of Payment.
This can be further configured according to the needs of the apartment such as taking in flat number or phone number.